requirements.txt
celery==4.4.0rc2
cx-Oracle==7.1.3
django
django-celery==3.3.0
django-filter==2.1.0
django-guardian==2.0.0
django-rest-framework-mongoengine==3.3.1
django-restframework==0.0.1
django-silk==3.0.2
django-timezone-field==3.0
djangorestframework==3.9.4
eventlet==0.25.0
mongoengine==0.17.0
mysqlclient==1.4.2.post1
psycopg2==2.8.2
pymongo==3.8.0
PyMySQL==0.9.3
redis==3.2.1
requests==2.21.0
Pillow
qiniu
anyjson
django-cors-headers
werkzeug
uwsgi
grequests



